Hand Glove Controlled Wheel Chair
This project is designed for physically disable people to control the wheel chair with thair hand gesture moments. In this project an accelerometer sensor is used to detect the hand gesture moments and give analog signal to an arduino nano. The arduino nano will convert an analog signal into digital signal and supply power to the relay. The relay will supply power to the motor and motor will rotate according to the relay commends.